feat: add tests for checkbox control
Introduces use of snapshots for testing controls. Also creates separate test
directories for prop-controllers and controls.
