+++ .changeset/tame-pianos-flow.md
@@ -1,0 +1,5 @@
+---
+"@makeswift/runtime": patch
+---
+
+Adds rendering tests for the checkbox control
