+++ packages/runtime/src/next/components/tests/prop-controllers/page-prop-controller.tsx
@@ -11,17 +11,17 @@ import {
   type Value,
 } from '@makeswift/prop-controllers'
 
-import { type ElementData, type ComponentType } from '../../../state/react-page'
+import { type ElementData, type ComponentType } from '../../../../state/react-page'
 import { randomUUID } from 'crypto'
 
-import { Page } from '../page'
+import { Page } from '../../page'
 import { act } from 'react-dom/test-utils'
-import { ReactRuntimeProvider } from '../../context/react-runtime'
-import { ReactRuntime } from '../../../react'
+import { ReactRuntimeProvider } from '../../../context/react-runtime'
+import { ReactRuntime } from '../../../../react'
 import {
   createMakeswiftPageSnapshot,
   createRootComponent,
-} from '../../../utils/tests/element-data-test-test'
+} from '../../../../utils/tests/element-data-test-test'
 
 export const pagePropControllerTest = <
   P extends PropDef & ((options?: any) => any),
